+++ src/css/index.css
@@ -1,3 +1,4 @@
+/* Top bar */
 nav {
     display: flex;
     justify-content: flex-start;
