+++ src/html/index.html
@@ -35,9 +35,9 @@
             <label for="show-render-area-button">Show Render Area: </label>
             <button id="show-render-area-button" value="false">Show</button>
         </div>
-        <div class="generate-test-object">
-            <label for="generate-test-object-button">Generate New Test Object: </label>
-            <button id="generate-test-object-button">Generate</button>
+        <div class="generate-test-entity">
+            <label for="generate-test-entity-button">Generate New Test Entity: </label>
+            <button id="generate-test-entity-button">Generate</button>
         </div>
         <div class="reset-diagram">
             <label>Reset Current Diagram: </label>
